+++ src/FamilyHubs.Referral.Core/DistributedCache/IReferralCacheKeys.cs
@@ -1,0 +1,7 @@
+﻿
+namespace FamilyHubs.Referral.Core.DistributedCache;
+
+public interface IReferralCacheKeys
+{
+    string ProfessionalReferral { get; }
+}
