+++ src/FamilyHubs.Referral.Core/Models/ProfessionalReferralModel.cs
@@ -1,12 +1,13 @@
 ﻿
-namespace FamilyHubs.Referral.Web.Models;
+namespace FamilyHubs.Referral.Core.Models;
 
 public class ProfessionalReferralModel
 {
     public string ServiceId { get; set; } = string.Empty;
     public string ServiceName { get; set; } = string.Empty;
+
     /// <summary>
     /// Full name of the contact
     /// </summary>
     public string FullName { get; set; } = string.Empty;
-}B
+}
