+++ src/FamilyHubs.Referral.Web/Extensions/StringExtensions.cs
@@ -2,6 +2,7 @@
 
 public static class StringExtensions
 {
+    //todo: move to file in core?
     public static string ToSentenceCase(this string input)
     {
         if (input.Length < 1)
